Begin tools subpackage: extract widgets, build, goal tools
Introduce a tools/ subpackage. Tool handlers move out of server.py into
thematic modules and reference shared infrastructure (mcp, helpers) via the
server module object, so existing test monkeypatching is unaffected.
server.py re-exports the handlers for backward compatibility.

This commit moves the widgets, build, and goal/term_goal tools; the
remaining groups follow. No functional change.
